rename the Overview component
